# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Simanja (Sistem Management Jamaah)",
        "description": "This is developped using Swagger UI.  You can find\r\nout more about Swagger at\r\n[http://swagger.io](http://swagger.io) or on\r\n[irc.freenode.net, #swagger](http://swagger.io/irc/).",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "iqbalhanif313@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/setup/desa": {
            "get": {
                "tags": [
                    "setup/desa"
                ],
                "summary": "Show List Setup Desa",
                "operationId": "setup/desa",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "setup/desa"
                ],
                "summary": "Store Setup Desa",
                "operationId": "setup/desa/store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Store Setup Desa",
                                        "type": "string",
                                        "example": "SPR"
                                    },
                                    "nama": {
                                        "description": "Store Setup Desa",
                                        "type": "string",
                                        "example": "Semampir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/desa/{id}": {
            "get": {
                "tags": [
                    "setup/desa"
                ],
                "summary": "Show Detail Setup Desa",
                "operationId": "setup/desa/{id}",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "put": {
                "tags": [
                    "setup/desa"
                ],
                "summary": "Update Setup Desa",
                "operationId": "setup/desa/update",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Update Setup Desa",
                                        "type": "string",
                                        "example": "SPR"
                                    },
                                    "nama": {
                                        "description": "Update Setup Desa",
                                        "type": "string",
                                        "example": "Semampir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            },
            "delete": {
                "tags": [
                    "setup/desa"
                ],
                "summary": "Delete Setup Desa",
                "operationId": "setup/desa/{id}/delete",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/jenis-kegiatan": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Jenis Kegiatan information",
                "operationId": "setup/jenis-kegiatan",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kabupaten": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Kabupaten information",
                "operationId": "setup/kabupaten",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
        "/api/setup/kabupaten/{st_provinsi_id}": {
            "get": {
                "tags": [
                    "setup"
                ],
                "summary": "Show Setup Kabupaten information",
                "operationId": "setup/kabupaten/id",
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "api_key": [
                            "write:user",
                            "read:user"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}